Fix the error in setting provision_cache_store fact
Filter 'default()' will work only for the failures in
the outer vars, if there are failures in case of inner vars
then default will not work, default has to be provided for the
inner var too ('registry_host' not available in 'groups').
Add conditionals to validated the inner vars and skip the task
to allow the default to be applied in following task, if undefined.