Add check to validate 'resources' presence in mcp query
After few (around 15) re-tries, when the master node is switching
during the update process, the query to get the resource object
resulted in a dict without 'resources' key, which should never
happen as all the resource query should have resource object
and the object static object (mcp master resource). This could
be because of the master node transition and the api server could
in the process of changing and the query has been initiated. Need
to check how ansible 'k8s_info' is handlig this scenarios.
For a workaround, added a check to valiated if 'resources' key
is present in the result object of the query.